aufgabe 10 und 11 alles fertig, endültig.

+++ src/main/java/at/spengergasse/spring_thymeleaf/entities/Reservation.java
@@ -9,7 +9,7 @@ public class Reservation
 {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     @ManyToOne(fetch = FetchType.EAGER)
     @JoinColumn(name = "patient_id")
@@ -38,12 +38,12 @@ public class Reservation
         return patient;
     }
 
-    public int getId()
+    public Long getId()
     {
         return id;
     }
 
-    public void setId(int id)
+    public void setId(Long id)
     {
         this.id = id;
     }
